moved main into cmd
